backend/requirements.txt
fastapi
uvicorn
pytest
pydantic
sqlmodel
google-generativeai
python-dotenv
